brain/requirements.txt
# Core dependencies
numpy>=1.24.0

# AI and ML
sentence-transformers>=2.2.0
transformers>=4.30.0
faiss-cpu>=1.7.0

# LLM APIs
openai>=1.0.0
google-generativeai>=0.3.0

# HTTP server
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0 